+++ src/test/java/com/dmz/airdnd/reservation/repository/ReservationRepositoryTest.java
@@ -11,8 +11,8 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import com.dmz.airdnd.AbstractContainerBase;
 import com.dmz.airdnd.accommodation.domain.Accommodation;
 import com.dmz.airdnd.accommodation.domain.Address;
-import com.dmz.airdnd.accommodation.domain.repository.AccommodationRepository;
-import com.dmz.airdnd.accommodation.domain.repository.AddressRepository;
+import com.dmz.airdnd.accommodation.repository.AccommodationRepository;
+import com.dmz.airdnd.accommodation.repository.AddressRepository;
 import com.dmz.airdnd.fixture.TestAccommodationFactory;
 import com.dmz.airdnd.fixture.TestAddressFactory;
 import com.dmz.airdnd.fixture.TestReservationFactory;
